refactor(audio): Wire the OpenAL backend in as a real audio device (#28)
The OpenAL Miles-API implementation was built only by a standalone CI harness. The
non-Windows branch of the root CMakeLists now supplies milesstub from it, so every
audio consumer links it without changing any of the 217 AIL_* call sites; 32-bit
Windows keeps the retail Miles path untouched.

Adds scripts/audio-surface-scan.py, which measures the demanded AIL_* surface and
classifies what the backend does with each entry point (50 implemented, 35 recorded,
6 no-op, 3 aliases, 0 missing; 101 declared == 101 defined). Corrects two numbers in
docs/porting/audio-surface.md and records the slice in docs/porting/audio-device-seam.md.

WWAudio's Win32 threading/event/timer code now uses the existing WWPlatform APIs off
Windows, and the audio callback payload widens to uintptr_t so pointers survive at
64-bit. Removes five includes of an osdep.h that does not exist in the repo.

+++ Core/Libraries/Source/WWVegas/WWAudio/SoundScene.cpp
@@ -199,7 +199,7 @@ SoundSceneClass::Collect_Logical_Sounds (unsigned int milliseconds, int listener
 				//	Is the sound ready to notify?
 				//
 				if (sound_obj->Allow_Notify (timestamp)) {
-					listener->On_Event (AudioCallbackClass::EVENT_LOGICAL_HEARD, (uint32)listener, (uint32)sound_obj);
+					listener->On_Event (AudioCallbackClass::EVENT_LOGICAL_HEARD, (uintptr_t)listener, (uintptr_t)sound_obj);
 				}
 			}
 		}
