+++ Core/Libraries/Source/WWVegas/WWAudio/SoundSceneObj.h
@@ -37,6 +37,7 @@
 #include "WWLib/always.h"
 #include "WWAudio.h"
 #include "WWLib/bittype.h"
+#include <Utility/stdint_adapter.h>
 #include "WWSaveLoad/persist.h"
 #include "WWLib/multilist.h"
 #include "WWLib/mutex.h"
@@ -123,7 +124,7 @@ class SoundSceneObjClass : public MultiListObjectClass, public PersistClass, pub
 		//////////////////////////////////////////////////////////////////////
 		//	Event handling
 		//////////////////////////////////////////////////////////////////////
-		virtual void			On_Event (AudioCallbackClass::EVENTS event, uint32 param1 = 0, uint32 param2 = 0);
+		virtual void			On_Event (AudioCallbackClass::EVENTS event, uintptr_t param1 = 0, uintptr_t param2 = 0);
 		virtual void			Register_Callback (AudioCallbackClass::EVENTS events, AudioCallbackClass *callback);
 
 		//////////////////////////////////////////////////////////////////////
@@ -225,8 +226,8 @@ __inline void
 SoundSceneObjClass::On_Event
 (
 	AudioCallbackClass::EVENTS	event,
-	uint32							param1,
-	uint32							param2
+	uintptr_t						param1,
+	uintptr_t						param2
 )
 {
 	if ((m_pCallback != nullptr) && (m_RegisteredEvents & event)) {
